Update README.md
# basic_python
This is a collection of basic python codes that will help at some point if you are using python for Data Science, Machine Learning and Deep Learning. The contents are in the form of code snippets.

Contents:
sklearn.utils.bunch to pandas Dataframe.ipynb - This is about how to convert a bunch object into a python DataFrame.